# openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Contact Directory API",
        "description": "REST API for managing contacts in the Contact Directory.",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
        "/api/v1/contacts/{id}": {
            "get": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Get a contact by id",
                "operationId": "get",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Contact identifier",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Contact not found"
                    },
                    "200": {
                        "description": "Contact found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ContactResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Replace a contact",
                "description": "Performs a full update of the contact with the given id.",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Contact identifier",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ContactRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "404": {
                        "description": "Contact not found"
                    },
                    "412": {
                        "description": "Stale version / concurrent modification"
                    },
                    "400": {
                        "description": "Validation failed"
                    },
                    "409": {
                        "description": "Email already exists"
                    },
                    "200": {
                        "description": "Contact updated",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ContactResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Soft-delete a contact",
                "description": "Soft-deletes the contact (moves it to trash); it can be restored or permanently deleted later.",
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Contact identifier",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Contact moved to trash"
                    },
                    "404": {
                        "description": "Contact not found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Partially update a contact",
                "description": "Updates only the supplied (non-null) fields of the contact.",
                "operationId": "patch",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Contact identifier",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ContactPatchRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "404": {
                        "description": "Contact not found"
                    },
                    "412": {
                        "description": "Stale version / concurrent modification"
                    },
                    "400": {
                        "description": "Validation failed"
                    },
                    "409": {
                        "description": "Email already exists"
                    },
                    "200": {
                        "description": "Contact updated",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ContactResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/contacts": {
            "get": {
                "tags": [
                    "Contacts"
                ],
                "summary": "List contacts",
                "description": "Returns a paginated list of contacts, optionally filtered by a search term and/or a tag.",
                "operationId": "list",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Free-text search across first name, last name, email, company and phone",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tag",
                        "in": "query",
                        "description": "Restrict results to contacts carrying this tag (case-insensitive)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Zero-based page index (0..N)",
                        "required": false,
                        "schema": {
                            "minimum": 0,
                            "type": "integer",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "description": "The size of the page to be returned",
                        "required": false,
                        "schema": {
                            "minimum": 1,
                            "type": "integer",
                            "default": 20
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            },
                            "default": [
                                "lastName,ASC"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Page of contacts",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageContactResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Create a contact",
                "description": "Creates a new contact and returns it with a Location header.",
                "operationId": "create",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ContactRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "400": {
                        "description": "Validation failed"
                    },
                    "409": {
                        "description": "Email already exists"
                    },
                    "201": {
                        "description": "Contact created",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ContactResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/contacts/bulk/tags": {
            "post": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Bulk add/remove tags on contacts",
                "description": "Adds and/or removes tags across the supplied contacts, skipping missing or already-deleted ids.",
                "operationId": "bulkTags",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BulkTagsRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Bulk tags result",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/BulkResult"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ContactRequest": {
                "required": [
                    "email",
                    "firstName",
                    "lastName"
                ],
                "type": "object",
                "properties": {
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phone": {
                        "pattern": "^[+0-9 ()\\-]{7,20}$",
                        "type": "string"
                    },
                    "company": {
                        "type": "string"
                    },
                    "tags": {
                        "uniqueItems": true,
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "favorite": {
                        "type": "boolean"
                    },
                    "notes": {
                        "type": "string"
                    },
                    "version": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
...
        }
    }
}